Refactor bot into modular cogs
